Add small fix to rectangle-line intersection test
To be frankly honest, I have no idea how it even worked in the first
place 🤯

The code previously only ever checked if the sign changed compared to
the rectangle's max corner. So if that corner was dragged over a link,
then the test would have worked as intended. Otherwise, not, and the
whole test would have returned false.