Add static attributes
Static attributes are attributes which don't have a pin and therefore
can't be linked to anything. Wrapping ImGui code in
BeginStaticAttribute/EndStaticAttribute means that you can still use
IsAttributeActive and IsAnyAttributeActive to detect if the attribute
was modified by the user.

This commit deprecates EndAttribute in favor of EndInputAttribute,
EndOutputAttribute, and EndStaticAttribute. There are now some
differences in the internal logic and so differentiating between them
just made the internal imnodes code nicer.