Add grid snapping, primary grid line coloration.
Grid snapping uses a simple nearest-point algorithm that works for negative
coordinates. It could be fancier, but that would involve snapping the
node's position independent of the actual origin or getting the absolute
mouse position and original click position in the node and generating the
snap from there.

Primary line drawing is very simple, just colors the origin lines a lighter
(or darker) shade. Could possibly be extended to draw every `n` lines
instead.